Build-on-Push: Add executables
The Jenkins jobs themselves contain as less logic as possible, so they
do not have to be updated when the logic changes. The jobs themselves
almost immediately call one of the scripts added in this commit (each
job type has a corresponding shell script). The shell scripts mostly
just prepares the environmen for the ci_commit executable, which handles
most of the high-level Build-on-Push logic.